backend/requirements.txt
fastapi==0.105.0
werkzeug==3.0.1
sqlalchemy==2.0.23
pydantic-settings==2.1.0
uvicorn==0.24.0.post1
psycopg2==2.9.9
